Tidied files and corrected name
